requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-excel==0.0.10
django-import-export==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
httpie==2.1.0
idna==2.9
jdcal==1.4.1
lml==0.0.9
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
pyexcel==0.6.1
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
